Rewrote the core WebResource class, with removed dead code, better content negotiation, including Datacite.
