+++ metrics/WebResource.py
@@ -1,536 +1,440 @@
-# import eventlet
-
-# from https://github.com/eventlet/eventlet/issues/670
-# eventlet.monkey_patch(select=False)
-# eventlet.monkey_patch()
-
-import time
-from ssl import SSLError
-from lxml import html
+import json
 import logging
+import os
+import uuid
+from typing import Dict, List, Optional, Sequence, Tuple
+from urllib.parse import urljoin
+
+import extruct
+import requests
+from lxml import html
+from rdflib import ConjunctiveGraph, URIRef
 from selenium import webdriver
-from selenium.webdriver.support.wait import WebDriverWait
-from webdriver_manager.chrome import ChromeDriverManager
 from selenium.webdriver.chrome.options import Options
 from selenium.webdriver.chrome.service import Service
-import extruct
-from pathlib import Path
-from rdflib import ConjunctiveGraph, URIRef, Namespace
-import requests
-import json
-import os
-import re
+from selenium.webdriver.support.ui import WebDriverWait
+from webdriver_manager.chrome import ChromeDriverManager
 
-from metrics.util import clean_kg_excluding_ns_prefix
+from metrics.util import clean_kg_excluding_ns_prefix, is_DOI, get_DOI
+
+logger = logging.getLogger(__name__)
 
 requests.packages.urllib3.disable_warnings(
     requests.packages.urllib3.exceptions.InsecureRequestWarning
 )
 
-logger = logging.getLogger("DEV") 
+# configure logger to print to console with a simple format, including line number
+logging.basicConfig(
+    level=logging.INFO,
+    format="%(asctime)s - %(levelname)s - %(name)s: %(lineno)d - %(message)s",
+)
 
 
 class WebResource:
-    prefs = {
-        "download_restrictions": 3,
-        "download.prompt_for_download": False,
-        "download.default_directory": "NUL",
+    COMMON_RDF_MIME_TYPES: Sequence[Tuple[str, str]] = (
+        ("application/ld+json", "json-ld"),
+        ("application/rdf+xml", "xml"),
+        ("text/turtle", "turtle"),
+        ("application/n-triples", "ntriples"),
+        ("text/n3", "n3"),
+        ("application/trig", "trig"),
+        ("application/n-quads", "nquads"),
+    )
+
+    RDF_MIME_TO_FORMAT: Dict[str, str] = {
+        "application/ld+json": "json-ld",
+        "application/rdf+xml": "xml",
+        "text/turtle": "turtle",
+        "application/n-triples": "ntriples",
+        "text/plain": "ntriples",
+        "text/n3": "n3",
+        "application/trig": "trig",
+        "application/n-quads": "nquads",
+        "text/x-nquads": "nquads",
     }
 
-    base_path = Path(__file__).parent.parent  # current directory
-
-    chrome_options = Options()
-    chrome_options.add_argument("--headless")
-    chrome_options.add_argument("--no-sandbox")
-    #chrome_options.add_argument("--disable-dev-shm-usage")
-    #chrome_options.add_argument("--disable-gpu")
-
-    proxy = os.getenv("HTTP_PROXY")
-    if proxy:
-        chrome_options.add_argument("--proxy-server=" + proxy)
-
-    try:
-        WEB_BROWSER_HEADLESS = webdriver.Chrome(
-            service=Service(ChromeDriverManager().install()), options=chrome_options
-        )
-        # WEB_BROWSER_HEADLESS.implicitly_wait(20)
-
-        SERVER_TIMEOUT = 10
-    except requests.ConnectionError as e:
-        print(e)
-        pass
-
-    status_code = None
-    content_type = None
-    mimetype = None
-    browser_selenium = None
-    html_selenium = None
-    html_requests = None
-    headers = None
-    links_headers = None
-
-    kg_var_strings = [
-        "kg_links_header",
-        "kg_auto",
-        "kg_brut",
-        "kg_links_html",
-        "kg_html",
-    ]
-
-    # source: https://docs.aws.amazon.com/neptune/latest/userguide/sparql-media-type-support.html
-    RDF_MEDIA_TYPES_MAPPING = {
-        "turtle": ["text/turtle"],
-        "xml": ["application/rdf+xml"],
-        "json-ld": ["application/ld+json"],
-        "ntriples": [
-            "application/n-triples",
-            "text/turtle",
-            "text/plain",
-        ],
-        "n3": ["text/n3"],
-        "trig": [
-            "application/trig",
-        ],
-        "trix": [
-            "application/trix",
-        ],
-        "nquads": ["application/n-quads", "text/x-nquads"],
+    LINK_RELATIONS_FOR_METADATA = {
+        "describedby",
+        "alternate",
+        "item",
+        "meta",
     }
 
-    def __init__(self, url, rdf_graph=None) -> None:
-        self.id = "WebResource Unique ID for cache"
+    def __init__(
+        self,
+        url: str,
+        rdf_graph: Optional[ConjunctiveGraph] = None,
+        timeout: int = 30,
+    ) -> None:
+        self.id = uuid.uuid4()
         self.url = url
+        self.timeout = timeout
 
-        # TODO rename variable
-        self.wr_dataset = ConjunctiveGraph()
+        self.status_code: Optional[int] = None
+        self.headers: Dict[str, str] = {}
+        self.content_type: Optional[str] = None
 
-        self.wr_dataset.namespace_manager.bind("sc", URIRef("http://schema.org/"))
-        self.wr_dataset.namespace_manager.bind("scs", URIRef("https://schema.org/"))
-        self.wr_dataset.namespace_manager.bind("bsc", URIRef("https://bioschemas.org/"))
-        self.wr_dataset.namespace_manager.bind(
-            "dct", URIRef("http://purl.org/dc/terms/")
-        )
+        self.dataset = ConjunctiveGraph()
+        self.dataset.namespace_manager.bind("sc", URIRef("http://schema.org/"))
+        self.dataset.namespace_manager.bind("scs", URIRef("https://schema.org/"))
+        self.dataset.namespace_manager.bind("bsc", URIRef("https://bioschemas.org/"))
+        self.dataset.namespace_manager.bind("dct", URIRef("http://purl.org/dc/terms/"))
 
-        self.init_kgs()
+        self.graph_uris = {
+            "content_neg": URIRef(f"{self.url}#content-neg"),
+            "mime_probe": URIRef(f"{self.url}#mime-probe"),
+            "datacite_probe": URIRef(f"{self.url}#datacite-probe"),
+            "html_jsonld": URIRef(f"{self.url}#html-jsonld"),
+            "html_rdfa": URIRef(f"{self.url}#html-rdfa"),
+            "html_microdata": URIRef(f"{self.url}#html-microdata"),
+        }
 
-        if rdf_graph is None:
-            response = requests.get(url)
-            self.headers = response.headers
-            self.status_code = response.status_code
-
-            # mimetypes from headers
-            mimetype = self.headers["Content-Type"].split(";")[0]
-            self.mimetype = mimetype
-
-            # list of possibles rdf formats from mimetypes
-            rdf_formats = self.get_rdf_format_from_contenttype(mimetype)
-            # print(self.headers)
-
-            cite_as, described_by, items = self.retrieve_links_from_headers()
-
-            # get RDF from HTTP headers
-            self.get_kg_from_header(described_by)
-
-            # if not html, try to retrieve rdf from possible rdf format
-            if mimetype != "text/html":
-                response = self.request_from_url(self.url)
-
-                # generate rdf graph from mapped mimetypes
-                for rdf_format in rdf_formats:
-                    # print(rdf_format)
-                    self.kg_auto = self.get_rdf_from_mimetype_match(
-                        self.url, rdf_format, self.kg_auto
-                    )
-
-                # if no rdf found: brute force testing each RDF formats regardless of mimetypes
-                if len(self.kg_auto) == 0:
-                    for rdf_format in self.RDF_MEDIA_TYPES_MAPPING.keys():
-                        self.kg_brut = self.get_rdf_from_mimetype_match(
-                            url, rdf_format, self.kg_brut
-                        )
-
-                logger.info(
-                    "Resource content_type is: " + self.headers["Content-Type"]
-                )
-
-            elif mimetype == "text/html":
-                self.html_content = self.get_html_selenium(url)
-
-                # should only use selenium here I think
-                # TODO get RDF from HTTP html
-                # response_request = self.request_from_url(self.url)
-                # but need request for status_code
-                # self.status_code = response_request.status_code
-                # self.html_requests = response_request.content
-
-                # if KG from HTTP link header == 0 look for link header in html
-                if self.kg_links_header == 0:
-                    links = self.retrieve_links_from_html(self.kg_html)
-
-                    for link in links:
-                        rel = link[0]
-                        if rel == "describedby":
-                            url = link[1]
-                            link_mimetype = link[2]
-                            rdf_formats = self.get_rdf_format_from_contenttype(
-                                link_mimetype
-                            )
-
-                            for rdf_format in rdf_formats:
-                                self.kg_links_html = self.get_rdf_from_mimetype_match(
-                                    url, rdf_format, self.kg_links_html
-                                )
-
-                # get RDF metadata directly from html content
-                self.html_to_rdf_extruct(self.html_content)
-
-            # get static RDF metadata (already available in html sources)
-            self.html_content = self.request_from_url(self.url)
-            self.html_to_rdf_extruct(self.html_content)
-
-            self.rdf = (
-                self.kg_auto
-                + self.kg_brut
-                + self.kg_links_header
-                + self.kg_links_html
-                + self.kg_html
-            )
-
-            self.kg_auto.bind("wr", Namespace("http://webresource/"))
-            # self.wr_dataset.add_graph(self.kg_auto)
-            for s, p, o in self.kg_auto:
-                self.wr_dataset.add((s, p, o, URIRef(self.url + "#mimetypes_match")))
-
-            self.kg_brut.bind("wr", Namespace("http://webresource/"))
-            # self.wr_dataset.add_graph(self.kg_brut)
-            for s, p, o in self.kg_brut:
-                self.wr_dataset.add((s, p, o, URIRef(self.url + "#rdfformats_match")))
-
-            self.kg_links_header.bind("wr", Namespace("http://webresource/"))
-            # self.wr_dataset.add_graph(self.kg_links_header)
-            for s, p, o in self.kg_links_header:
-                self.wr_dataset.add((s, p, o, URIRef(self.url + "#links_header")))
-
-            self.kg_links_html.bind("wr", Namespace("http://webresource/"))
-            # self.wr_dataset.add_graph(self.kg_links_html)
-            for s, p, o in self.kg_links_html:
-                self.wr_dataset.add((s, p, o, URIRef(self.url + "#links_html")))
-
-            self.kg_html.bind("wr", Namespace("http://webresource/"))
-            # self.wr_dataset.add_graph(self.kg_html)
-            for s, p, o in self.kg_html:
-                self.wr_dataset.add((s, p, o, URIRef(self.url + "#html")))
-
-            # define the correct namespace for sc prefix
-            # if self.is_schema_http(self.wr_dataset):
-            #     self.wr_dataset.namespace_manager.bind("sc", URIRef("http://schema.org/"))
-            # else:
-            #     self.wr_dataset.namespace_manager.bind("sc", URIRef("https://schema.org/"))
-
-            # replace all http with https for schema.org
-            # self.wr_dataset.namespace_manager.bind("sc", URIRef("https://schema.org/"))
-            # self.wr_dataset = self.schema_https_convert(self.wr_dataset)
-
-            self.wr_dataset = clean_kg_excluding_ns_prefix(
-                self.wr_dataset, "http://www.w3.org/1999/xhtml/vocab#"
-            )
-
-            # print("HTML: " + str(len(self.kg_html)))
-            # print("LINKS HEADERS: " + str(len(self.kg_links_header)))
-            # print("AUTO: " + str(len(self.kg_auto)))
-            # print("FORMATS_GUESSING: " + str(len(self.kg_brut)))
-            # print("HTML LINKS: " + str(len(self.kg_links_html)))
-
+        if rdf_graph is not None:
+            for s, p, o in rdf_graph:
+                self.dataset.add((s, p, o, URIRef(f"{self.url}#provided")))
         else:
-            self.rdf = rdf_graph
+            self._retrieve_all_metadata()
 
-        # self.rdf.namespace_manager.bind("sc", URIRef("http://schema.org/"))
-        self.rdf.namespace_manager.bind("bsc", URIRef("https://bioschemas.org/"))
-        self.rdf.namespace_manager.bind("dct", URIRef("http://purl.org/dc/terms/"))
-        self.rdf = clean_kg_excluding_ns_prefix(
-            self.rdf, "http://www.w3.org/1999/xhtml/vocab#"
+        # remove triples with the xhtml vocab namespace,
+        # as they are often noise in this context and not relevant for FAIR assessment
+        self.dataset = clean_kg_excluding_ns_prefix(
+            self.dataset, "http://www.w3.org/1999/xhtml/vocab#"
         )
 
-        logger.info(f"Web resource {self.url} loaded with {len(self.rdf)} RDF triples")
+        logger.info(
+            "WebResourceV2 loaded %s with %s RDF triples",
+            self.url,
+            len(self.dataset),
+        )
 
-    def init_kgs(self):
-        for var_str in self.kg_var_strings:
-            setattr(
-                WebResource,
-                var_str,
-                ConjunctiveGraph(identifier="http://webresource/" + var_str),
-            )
-            getattr(WebResource, var_str).namespace_manager.bind(
-                "bsc", URIRef("https://bioschemas.org/")
-            )
-            getattr(WebResource, var_str).namespace_manager.bind(
-                "dct", URIRef("http://purl.org/dc/terms/")
-            )
-            clean_kg_excluding_ns_prefix(
-                getattr(WebResource, var_str), "http://www.w3.org/1999/xhtml/vocab#"
-            )
-
-    def get_kg_from_header(self, described_by):
-        for link in described_by:
-            reg_string = '<(.*?)>*;*rel="(.*?)"*;*type="(.*?)"'
-            match = re.search(reg_string, link)
-            url = match.group(1)
-            link_mimetype = match.group(3)
-
-            rdf_formats = self.get_rdf_format_from_contenttype(link_mimetype)
-
-            for rdf_format in rdf_formats:
-                self.get_rdf_from_mimetype_match(url, rdf_format, self.kg_links_header)
-
-    def get_url(self):
+    def get_url(self) -> str:
         return self.url
 
-    def get_rdf(self):
-        return self.wr_dataset
+    def get_rdf(self) -> ConjunctiveGraph:
+        return self.dataset
 
-    def get_wr_kg_dataset(self):
-        return self.wr_dataset
-
-    def get_kg_auto(self):
-        return self.kg_auto
-
-    def get_kg_html(self):
-        return self.kg_html
-
-    def get_status_code(self):
+    def get_status_code(self) -> Optional[int]:
         return self.status_code
 
-    def get_html_requests(self):
-        return self.html_requests
-
-    def get_http_header(self):
+    def get_http_header(self) -> Dict[str, str]:
         return self.headers
 
-    @staticmethod
-    def is_schema_http(kg):
-        for s, p, o, g in kg.quads(None):
-            if str(s).startswith("http://schema.org"):
-                return True
-            if str(p).startswith("http://schema.org"):
-                return True
-            if isinstance(o, URIRef):
-                if str(o).startswith("http://schema.org"):
-                    return True
-        return False
+    def _retrieve_all_metadata(self) -> None:
+        base_response = self._http_get(self.url)
+        if base_response is None:
+            return
 
-    # Could be a static method
-    def schema_https_convert(self, kg):
-        for s, p, o, g in kg.quads(None):
-            changed = False
-            new_s = s
-            if str(s).startswith("http://schema.org"):
-                new_s = URIRef(str(s).replace("http", "https", 1))
-                changed = True
-            new_p = p
-            if str(p).startswith("http://schema.org"):
-                new_p = URIRef(str(p).replace("http", "https", 1))
-                changed = True
-            new_o = o
-            if isinstance(o, URIRef):
-                if str(o).startswith("http://schema.org"):
-                    new_o = URIRef(str(o).replace("http", "https", 1))
-                    changed = True
-            if changed:
-                kg.remove((s, p, o, g))
-                kg.add((new_s, new_p, new_o, g))
-        return kg
-
-    def get_rdf_from_mimetype_match(self, url, rdf_format, kg):
-        logging.debug("Getting RDF from: " + rdf_format)
-
-        kg_temp = ConjunctiveGraph()
-        response = requests.get(url)
-
-        if response.status_code == 200:
-            rdf_str = response.text
-            try:
-                # TODO check if merging works with ConjunctiveGraph using publicID (seems to keep only latest)
-                kg_temp.parse(
-                    data=rdf_str,
-                    format=rdf_format,
-                    publicID=url,
-                )
-
-                kg += kg_temp
-                logging.debug(len(kg_temp))
-            except Exception as err:
-                # if error UnicodeDecodeError execute following code, otherwise continue to next format
-                if type(err).__name__ == "BadSyntax":
-                    print("RDF syntax error")
-                    print(err)
-                if type(err).__name__ == "UnicodeDecodeError":
-                    print("UNICODE error")
-                    print(err)
-                    kg = self.handle_unicodedecodeerror(url, kg, response)
-
-        return kg
-
-    def get_rdf_format_from_contenttype(self, mimetype):
-        # all_mediatypes = [
-        #     item
-        #     for sublist in self.RDF_MEDIA_TYPES_MAPPING.values()
-        #     for item in sublist
-        # ]
-        rdf_formats = [
-            i
-            for i in self.RDF_MEDIA_TYPES_MAPPING
-            if mimetype in self.RDF_MEDIA_TYPES_MAPPING[i]
-        ]
-
-        return rdf_formats
-
-    def handle_unicodedecodeerror(self, url, kg, response):
-        print("Handle JSON-LD parsing error")
-        # base_path = Path(__file__).parent.parent  # current directory
-        # static_file_path = str((base_path / "static/data/jsonldcontext.json").resolve())
-        json_response = response.json()
-
-        if "@context" in json_response.keys():
-            if ("https://schema.org" in json_response["@context"]) or (
-                "http://schema.org" in json_response["@context"]
-            ):
-                json_response["@context"] = self.static_file_path
-
-        json_str = json.dumps(json_response, ensure_ascii=False)
-
-        kg_temp = ConjunctiveGraph()
-
-        kg_temp.parse(data=json_str, format="json-ld", publicID=url)
-        for s, p, o in kg_temp:
-            kg.add((s, p, o))
-
-        return kg
-
-    def retrieve_links_from_headers(self):
-        links_col = []
-        decsribed_by_col = []
-        cite_as_col = []
-        item_col = []
-        headers = self.headers
-        for k in headers.keys():
-            if "link" in k.lower():
-                l_header = headers[k]
-                links = l_header.split(",")
-                for link in links:
-                    # print("----")
-                    links_col.append(link)
-                    tokens = link.split(";")
-                    # print(tokens)
-                    for t in tokens:
-                        if 'rel="describedby"' in t:
-                            decsribed_by_col.append(link)
-                        elif 'rel="item"' in t:
-                            item_col.append(link)
-                        elif 'rel="cite-as"' in t:
-                            cite_as_col.append(link)
-        self.links_headers = (cite_as_col, decsribed_by_col, item_col)
-        return cite_as_col, decsribed_by_col, item_col
-
-    def retrieve_links_from_html(self, source_html):
-        tree = html.fromstring(source_html)
-        links = []
-
-        txt = tree.xpath('//link[contains(@rel, "describedby")]')
-        for t in txt:
-            links.append(t.values())
-
-        txt = tree.xpath('//link[contains(@rel, "cite-as")]')
-        for t in txt:
-            links.append(t.values())
-
-        txt = tree.xpath('//link[contains(@rel, "item")]')
-        for t in txt:
-            links.append(t.values())
-        return links
-
-    # TODO Extruct can work with Selenium
-
-    def request_from_url(self, url):
-        nb_retry = 0
-        while nb_retry < 3:
-            try:
-                nb_retry += 1
-                response = requests.get(url=url, timeout=30, verify=False)
-                break
-            except SSLError:
-                time.sleep(5)
-            except requests.exceptions.Timeout:
-                print("Timeout, retrying")
-                time.sleep(5)
-            except requests.exceptions.ConnectionError as e:
-                print(e)
-                print("ConnectionError, retrying...")
-                time.sleep(10)
-
-        return response.text
-
-    def get_html_selenium(self, url):
-        #try:
-        browser = WebResource.WEB_BROWSER_HEADLESS
-        browser.get(url)
-        #time.sleep(2)
-        browser.set_page_load_timeout(30)
-        browser.implicitly_wait(30)
-        WebDriverWait(self.WEB_BROWSER_HEADLESS, self.SERVER_TIMEOUT).until(
-            lambda wd: self.WEB_BROWSER_HEADLESS.execute_script(
-                    "return document.readyState"
-                )
-                == "complete",
-                "Page taking too long to load",
-            )
-        html_content = browser.page_source
-        logging.debug(type(browser.page_source))
-        logging.info(f"size of the parsed web page: {len(html_content)}")
-        #finally:
-        #    browser.quit()
-        
-        return html_content
-
-    # @staticmethod
-    def html_to_rdf_extruct(self, html_source) -> ConjunctiveGraph:
-        data = extruct.extract(
-            html_source, syntaxes=["microdata", "rdfa", "json-ld"], errors="ignore"
+        self.status_code = base_response.status_code
+        self.headers = dict(base_response.headers)
+        self.content_type = self._normalize_content_type(
+            base_response.headers.get("Content-Type")
         )
 
-        kg_jsonld = ConjunctiveGraph()
-        if "json-ld" in data.keys():
-            for md in data["json-ld"]:
-                kg_jsonld += kg_jsonld.parse(
-                    data=json.dumps(md, ensure_ascii=False),
+        # self._collect_from_link_relations(base_response)
+        self._collect_from_common_accept_headers()
+        g1 = self.dataset.get_context(self.graph_uris["mime_probe"])
+
+        # if DOI, try datacite metadata retrieval
+        if is_DOI(self.url):
+            self._collect_from_datacite()
+
+        # if no triples were retrieved by content negotiation, try to collect embedded RDF with Selenium and extruct (costly)
+        if len(g1) == 0:
+            self._collect_embedded_rdf_with_selenium()
+
+    def _http_get(
+        self,
+        url: str,
+        headers: Optional[Dict[str, str]] = None,
+    ) -> Optional[requests.Response]:
+        try:
+            response = requests.get(
+                url,
+                headers=headers,
+                timeout=self.timeout,
+                verify=False,
+            )
+            return response
+        except requests.RequestException as exc:
+            logger.warning("HTTP request failed for %s: %s", url, exc)
+            return None
+
+    def _collect_from_link_relations(self, base_response: requests.Response) -> None:
+        graph = self.dataset.get_context(self.graph_uris["content_neg"])
+        candidates = self._extract_link_candidates(base_response)
+
+        for link_url, mime_hint in candidates:
+            formats = self._formats_for_mime(mime_hint)
+            if not formats:
+                formats = [fmt for _, fmt in self.COMMON_RDF_MIME_TYPES]
+
+            response = self._http_get(link_url)
+            if response is None or response.status_code >= 400:
+                continue
+
+            if not self._parse_response_in_formats(
+                graph,
+                source_url=link_url,
+                response=response,
+                candidate_formats=formats,
+            ):
+                inferred_format = self._format_from_response_content_type(response)
+                if inferred_format:
+                    self._parse_response_in_formats(
+                        graph,
+                        source_url=link_url,
+                        response=response,
+                        candidate_formats=[inferred_format],
+                    )
+
+    def _extract_link_candidates(
+        self, base_response: requests.Response
+    ) -> List[Tuple[str, Optional[str]]]:
+        candidates: List[Tuple[str, Optional[str]]] = []
+        seen = set()
+
+        raw_link_header = base_response.headers.get("Link")
+        if raw_link_header:
+            links = requests.utils.parse_header_links(raw_link_header)
+            for link in links:
+                href = link.get("url")
+                rel = (link.get("rel") or "").lower()
+                mime_hint = link.get("type")
+
+                if not href:
+                    continue
+                rel_tokens = {
+                    token.strip() for token in rel.split(" ") if token.strip()
+                }
+                if not rel_tokens.intersection(self.LINK_RELATIONS_FOR_METADATA):
+                    continue
+
+                absolute_href = urljoin(self.url, href)
+                key = (absolute_href, mime_hint)
+                if key not in seen:
+                    seen.add(key)
+                    candidates.append(key)
+
+        if (
+            self._normalize_content_type(base_response.headers.get("Content-Type"))
+            == "text/html"
+        ):
+            try:
+                doc = html.fromstring(base_response.text)
+                for node in doc.xpath("//link[@href]"):
+                    rel = (node.get("rel") or "").lower()
+                    rel_tokens = {
+                        token.strip()
+                        for token in rel.replace(",", " ").split(" ")
+                        if token.strip()
+                    }
+                    if not rel_tokens.intersection(self.LINK_RELATIONS_FOR_METADATA):
+                        continue
+
+                    href = node.get("href")
+                    if not href:
+                        continue
+                    mime_hint = node.get("type")
+                    absolute_href = urljoin(self.url, href)
+                    key = (absolute_href, mime_hint)
+                    if key not in seen:
+                        seen.add(key)
+                        candidates.append(key)
+            except Exception as exc:
+                logger.debug("Could not parse HTML links for %s: %s", self.url, exc)
+
+        return candidates
+
+    def _collect_from_datacite(self) -> None:
+        graph = self.dataset.get_context(self.graph_uris["datacite_probe"])
+        doi = get_DOI(self.url)
+        if not doi:
+            return
+
+        logger.info(f"Collecting metadata from datacite for DOI {doi}")
+        datacite_endpoint = f"https://data.crosscite.org/application/ld+json/{doi}"
+        response = requests.get(datacite_endpoint)
+        if response.status_code >= 400:
+            logger.warning(
+                "HTTP request failed for %s: %s", datacite_endpoint, response
+            )
+            return
+        else:
+            metadata = response.text
+            kg = ConjunctiveGraph()
+            try:
+                kg.parse(data=metadata, format="json-ld")
+            except Exception as e:
+                logger.warning("Could not parse metadata from datacite")
+                logger.warning(e)
+                return
+
+            if len(kg) > 0:
+                logger.info(f"Collected {len(kg)} triples from datacite for {doi}")
+                # logger.info(kg.serialize(format="turtle"))
+            # merge graph and kg
+            graph += kg
+
+    def _collect_from_common_accept_headers(self) -> None:
+        graph = self.dataset.get_context(self.graph_uris["mime_probe"])
+
+        for accept_mime, rdf_format in self.COMMON_RDF_MIME_TYPES:
+            response = self._http_get(self.url, headers={"Accept": accept_mime})
+            if response is None or response.status_code >= 400:
+                continue
+
+            parsed = self._parse_response_in_formats(
+                graph,
+                source_url=self.url,
+                response=response,
+                candidate_formats=[rdf_format],
+            )
+
+            if parsed:
+                logger.info(
+                    f"Parsed RDF triples: {len(graph)} triples for {accept_mime}: {rdf_format}"
+                )
+            else:
+                inferred_format = self._format_from_response_content_type(response)
+                if inferred_format:
+                    self._parse_response_in_formats(
+                        graph,
+                        source_url=self.url,
+                        response=response,
+                        candidate_formats=[inferred_format],
+                    )
+                    logger.info(
+                        f"Parsed RDF triples: {len(graph)} triples for {accept_mime} with content-type inference: {inferred_format}"
+                    )
+
+    def _collect_embedded_rdf_with_selenium(self) -> None:
+        logger.info("Collecting embedded RDF")
+        driver = None
+        service = None
+
+        try:
+            chrome_options = Options()
+            chrome_options.add_argument("--headless=new")
+            chrome_options.add_argument("--no-sandbox")
+            chrome_options.add_argument("--disable-dev-shm-usage")
+
+            proxy = os.getenv("HTTP_PROXY")
+            if proxy:
+                chrome_options.add_argument("--proxy-server=" + proxy)
+
+            service = Service(ChromeDriverManager().install())
+            driver = webdriver.Chrome(service=service, options=chrome_options)
+            driver.set_page_load_timeout(self.timeout)
+            driver.get(self.url)
+            WebDriverWait(driver, self.timeout).until(
+                lambda d: d.execute_script("return document.readyState") == "complete"
+            )
+            html_source = driver.page_source
+
+            data = extruct.extract(
+                html_source,
+                base_url=self.url,
+                syntaxes=["json-ld", "rdfa", "microdata"],
+                errors="ignore",
+            )
+
+            self._parse_extruct_json_items(data.get("json-ld", []), "html_jsonld")
+            self._parse_extruct_json_items(data.get("rdfa", []), "html_rdfa")
+            self._parse_extruct_json_items(data.get("microdata", []), "html_microdata")
+        except Exception as exc:
+            logger.warning("Selenium extraction failed for %s: %s", self.url, exc)
+        finally:
+            # Always close webdriver and service to avoid orphan chromedriver processes.
+            if driver is not None:
+                try:
+                    driver.quit()
+                except Exception as exc:
+                    logger.debug(
+                        "Error while quitting webdriver for %s: %s", self.url, exc
+                    )
+            if service is not None:
+                try:
+                    service.stop()
+                except Exception as exc:
+                    logger.debug(
+                        "Error while stopping chromedriver service for %s: %s",
+                        self.url,
+                        exc,
+                    )
+
+    def _parse_extruct_json_items(self, items: Sequence[dict], graph_key: str) -> None:
+        graph = self.dataset.get_context(self.graph_uris[graph_key])
+        for item in items:
+            try:
+                graph.parse(
+                    data=json.dumps(item, ensure_ascii=False),
                     format="json-ld",
                     publicID=self.url,
                 )
+            except Exception:
+                continue
 
-        kg_rdfa = ConjunctiveGraph()
-        if "rdfa" in data.keys():
-            for md in data["rdfa"]:
-                kg_rdfa += kg_rdfa.parse(
-                    data=json.dumps(md, ensure_ascii=False),
-                    format="json-ld",
-                    publicID=self.url,
-                )
+    def _parse_response_in_formats(
+        self,
+        graph,
+        source_url: str,
+        response: requests.Response,
+        candidate_formats: Sequence[str],
+    ) -> bool:
+        body_text = response.text
+        parsed_any = False
 
-        kg_microdata = ConjunctiveGraph()
-        if "microdata" in data.keys():
-            for md in data["microdata"]:
-                kg_microdata += kg_microdata.parse(
-                    data=json.dumps(md, ensure_ascii=False),
-                    format="json-ld",
-                    publicID=self.url,
-                )
+        for rdf_format in candidate_formats:
+            kg = ConjunctiveGraph()
+            try:
+                kg.parse(data=body_text, format=rdf_format, publicID=source_url)
+                parsed_any = True
+            except Exception:
+                continue
 
-        kg_extruct = kg_jsonld + kg_rdfa + kg_microdata
+            # merge graph and kg
+            graph += kg
 
-        for s, p, o in kg_extruct:
-            self.kg_html.add((s, p, o))
+        return parsed_any
+
+    def _format_from_response_content_type(
+        self, response: requests.Response
+    ) -> Optional[str]:
+        content_type = self._normalize_content_type(
+            response.headers.get("Content-Type")
+        )
+        return self.RDF_MIME_TO_FORMAT.get(content_type)
+
+    def _formats_for_mime(self, mime: Optional[str]) -> List[str]:
+        if not mime:
+            return []
+        normalized = self._normalize_content_type(mime)
+        rdf_format = self.RDF_MIME_TO_FORMAT.get(normalized)
+        return [rdf_format] if rdf_format else []
+
+    @staticmethod
+    def _normalize_content_type(content_type: Optional[str]) -> str:
+        if not content_type:
+            return ""
+        return content_type.split(";", 1)[0].strip().lower()
+
+    # def _is_html_only(self, base_response: requests.Response) -> bool:
+    #     is_html = (
+    #         self._normalize_content_type(base_response.headers.get("Content-Type"))
+    #         == "text/html"
+    #     )
+    #     if not is_html:
+    #         return False
+    #
+    #     content_neg_size = len(self.dataset.get_context(self.graph_uris["content_neg"]))
+    #     mime_probe_size = len(self.dataset.get_context(self.graph_uris["mime_probe"]))
+    #     return (content_neg_size + mime_probe_size) == 0
 
     def __str__(self) -> str:
-        out = """Web resource under FAIR assesment:\n\t"""
-        out += self.url + "\n\t"
-        out += str(len(self.rdf)) + " embedded RDF triples"
-        return out
+        return (
+            "Web resource under FAIR assessment:\n\t"
+            + self.url
+            + "\n\t"
+            + str(len(self.dataset))
+            + " RDF triples"
+        )
